finished with the code, going to refactor to see if I can extend code instead for cleaner
