+++ project/Assets/Scenes/SampleScene.unity
@@ -1226,6 +1226,31 @@ PrefabInstance:
       propertyPath: obj
       value: 
       objectReference: {fileID: 2099042040}
+    - target: {fileID: 7994346040598928315, guid: 46b097daf62b21c409354716e8730cc8,
+        type: 3}
+      propertyPath: target
+      value: 
+      objectReference: {fileID: 2099042040}
+    - target: {fileID: 7994346040598928315, guid: 46b097daf62b21c409354716e8730cc8,
+        type: 3}
+      propertyPath: speedLimit
+      value: 8
+      objectReference: {fileID: 0}
+    - target: {fileID: 7994346040598928315, guid: 46b097daf62b21c409354716e8730cc8,
+        type: 3}
+      propertyPath: isTerminating
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 7994346040598928315, guid: 46b097daf62b21c409354716e8730cc8,
+        type: 3}
+      propertyPath: isTouch
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 7994346040598928315, guid: 46b097daf62b21c409354716e8730cc8,
+        type: 3}
+      propertyPath: nRotationsToFinish
+      value: 5
+      objectReference: {fileID: 0}
     m_RemovedComponents:
     - {fileID: 1377129301697488976, guid: 46b097daf62b21c409354716e8730cc8, type: 3}
   m_SourcePrefab: {fileID: 100100000, guid: 46b097daf62b21c409354716e8730cc8, type: 3}
