--- project/Assets/Scripts/Rotation/MouseRotator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 34cf6a58335094b9abd8e684d1a8475a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
