+++ project/Assets/Scripts/Rotation/NewMonoBehaviour.cs
@@ -1,0 +1,17 @@
+﻿using UnityEngine;
+using System.Collections;
+
+public class NewMonoBehaviour : MonoBehaviour
+{
+    // Use this for initialization
+    void Start()
+    {
+
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+
+    }
+}
