+++ project/Assets/Scripts/Rotation/ObjectRotator.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c1a3edec4dff4f441a00cfab0dde3763
+guid: 34cf6a58335094b9abd8e684d1a8475a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
