--- project/Assets/Scripts/Rotation/WheelRotator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 839410cea4ecddc4cb4e2e151afbee30
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
